// src/models/SocietyChargeHeadsModel.tsx
export interface SocietyChargeHeadsModel {

    SocietyId: string;
    ChargeHeadId: string;
    ChargeHead: string;
    ChargeInterest: boolean;
    AcHeadId: string;
    Nature: string;
    Rate: number;
    Sequence: number;
    BillAbbreviation: string;
    ChargeTax: boolean;
    Hsncode :string;
    MonthlyCharge : boolean;
}